tests/wpt/web-platform-tests/html/browsers/browsing-the-web/unloading-documents/002.html

<!doctype html>
<title>document.open in unload</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>
<script>
var t = async_test();
var win;

t.step(function() {
  win = window.open("002-1.html");
});

add_completion_callback(function() {win.close()});
</script>
