third_party/blink/web_tests/external/wpt/html/browsers/browsing-the-web/unloading-documents/002-expected.txt

This is a testharness.js-based test.
[FAIL] document.open in unload
  assert_equals: expected "016789Z" but got "01689Z"
Harness: the test ran to completion.
